+++ src/main/java/org/folio/inventory/dataimport/exceptions/InstanceLinksException.java
@@ -1,0 +1,9 @@
+package org.folio.inventory.dataimport.exceptions;
+
+public class InstanceLinksException extends RuntimeException {
+
+  public InstanceLinksException(String message) {
+    super(message);
+  }
+}
+